// types.ts
export interface BashSnippet {
  id: string;
  nameKey: string;
  descriptionKey: string;
  content: string;
}